requirements.txt
psycopg2==2.6.2
wiki==0.2b2
django-extensions==1.7.4
